feat: extract and implement auth module
