+++ src/api/cleo-api.ts
@@ -4,12 +4,13 @@ import {corsOptions} from "./cors-config";
 const cors = require('cors');
 import {sessionMiddleware} from "./session-config";
 import passport = require("passport");
-require("../modules/user/passport-config");
+require("../modules/auth/passport-config");
 import journalsRouter from "../modules/journal/journals-router";
 import entriesRouter from "../modules/entry/entries-router";
 import {API_PORT, API_TITLE} from "../environment";
 import usersRouter from "../modules/user/users-router";
-import {authGuard} from "../modules/user/users-request-handlers";
+import authRouter from "../modules/auth/auth-router";
+import {authGuard} from "../modules/auth/auth-request-handlers";
 
 const cleoHomeRoute = (req, res): RequestHandler =>
     res.send(API_TITLE);
@@ -22,7 +23,8 @@ app.use(passport.initialize());
 app.use(passport.session());
 
 app.get('/', cleoHomeRoute);
-app.use('/users/', usersRouter);
+app.use('/auth/', authRouter);
+app.use('/users/', authGuard, usersRouter);
 app.use('/journals/', authGuard, journalsRouter);
 app.use('/entries/', authGuard, entriesRouter);
