+++ src/modules/auth/passport-config.ts
@@ -1,7 +1,7 @@
-import {GetUserDTO} from "./users-dtos";
+import {GetUserDTO} from "../user/users-dtos";
 const passportConfig = require('passport');
 import LocalStrategy = require('passport-local');
-import {User} from "./user";
+import {User} from "../user/user";
 import {USERS_REPOSITORY} from "../../repositories-config";
 import {validateHash} from "../../utils/password-utils";
