+++ src/modules/user/users-request-handlers.ts
@@ -1,12 +1,10 @@
-import {NextFunction, Request, RequestHandler, Response} from "express";
+import {Request, RequestHandler, Response} from "express";
 import {User, UserSortOption, UserStatusOption} from "./user";
 import {
     deleteUser,
     deleteUsers,
     getUser,
     getUsers,
-    registerAdminUser,
-    registerUser,
     updateUser,
     updateUsers
 } from "./users-service";
@@ -15,55 +13,11 @@ import {
     DeleteUsersDTO,
     GetUserDTO,
     GetUsersDTO,
-    RegisterAdminDTO,
-    RegisterUserDTO,
     UpdateUserDTO,
     UpdateUsersDTO
 } from "./users-dtos";
 import {sendErrorResponse} from "../../utils/send-error-response";
-import {
-    CREATED,
-    INTERNAL_SERVER_ERROR,
-    UNAUTHORIZED
-} from "../../utils/http-status-constants";
 import {OrderOption} from "../../utils/order-option";
-import passport = require("passport");
-
-export const authenticate: RequestHandler = passport.authenticate('local');
-
-export const authGuard: RequestHandler = (req: Request, res: Response, next: NextFunction) => {
-    if (!req.isAuthenticated())
-        return res.status(UNAUTHORIZED).json('Unauthorized.');
-    return next();
-};
-
-export const register: RequestHandler = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
-    try {
-        const dto: RegisterUserDTO = mapToRegisterUserDTO(req);
-        await registerUser(dto);
-        return next();
-    } catch (error) {
-        sendErrorResponse(error, res);
-    }
-};
-
-export const registerAdmin: RequestHandler = async (req: Request, res: Response): Promise<void> => {
-    try {
-        const dto: RegisterAdminDTO = mapToRegisterAdminDTO(req);
-        const user = await registerAdminUser(req.user as User, dto);
-        res.status(CREATED).json(user);
-    } catch (error) {
-        sendErrorResponse(error, res);
-    }
-};
-
-export const authenticatedResponse: RequestHandler = (req: Request, res: Response): void => {
-    res.json({"username": (req.user as User).username});
-};
-
-export const loggedInResponse: RequestHandler = (req: Request, res: Response): void => {
-    res.json({"username": (req.user as User).username});
-};
 
 export const getUsersHandler: RequestHandler = async (req: Request, res: Response) => {
     try {
@@ -123,26 +77,6 @@ export const deleteUserHandler: RequestHandler = async (req: Request, res: Respo
         sendErrorResponse(error, res);
     }
 };
-
-export const logout: RequestHandler = (req: Request, res: Response): void => {
-    req.logout((error) => {
-        if (error) {
-            res.status(INTERNAL_SERVER_ERROR).json('Log out failed.');
-            return;
-        }
-        res.json('Logged out successfully.');
-    });
-};
-
-const mapToRegisterUserDTO = (req: Request): RegisterUserDTO => ({
-    username: req.body.username,
-    password: req.body.password,
-});
-
-const mapToRegisterAdminDTO = (req: Request): RegisterAdminDTO => ({
-    username: req.body.username,
-    password: req.body.password,
-});
 
 const mapToGetUsersDTO = (req: Request): GetUsersDTO => ({
     ... req.query.username && {username: req.query.username as string},
