+++ src/modules/user/users-router.ts
@@ -4,23 +4,11 @@ import {
     getUsersHandler,
     deleteUserHandler,
     updateUserHandler,
-    authenticate,
-    loggedInResponse,
-    authenticatedResponse,
-    logout,
-    register,
-    authGuard,
-    registerAdmin,
     updateUsersHandler,
     deleteUsersHandler,
 } from "./users-request-handlers";
 
 const usersRouter: Router = Router();
-usersRouter.post('/login/', authenticate, loggedInResponse);
-usersRouter.post('/logout/', authGuard, logout);
-usersRouter.post('/register/', register, authenticate, loggedInResponse);
-usersRouter.post('/register-admin/', registerAdmin);
-usersRouter.get('/protected/', authGuard, authenticatedResponse);
 
 usersRouter.get('/', getUsersHandler);
 usersRouter.get('/:username', getUserHandler);
